Made separate folders for views (html) and scripts, also added in external graphics library fabric.js which is included in the scripts directory
Added bootstrap for styling. Both the bootstrap and fabric.js libraries are available via bower, a package manager for web stuff, which we can start using if people want. I also added in a new view that has a canvas and uses fabric.js. Fabric.js automatically seems to take care of rotations, resizing and moving objects in an html canvas for basic shapes and any images that we add. I am tentatively pushing this, but we can remove it and talk about what libraries we want to use and what not on Wednesday.

Added functionality to add and delete generic rectangle in draw.html